// src/renderer/views/page/DataEntry/KojiMaking/entity/Enum.ts
export enum EMPType {
  TEMP = '临时工',
  FORMAL = '正式',
  TRANSFER = '借调'
}
export enum DayType {
  DAY = '白班',
  MIDDLE = '中班',
  NIGHT = '夜班'
}
export enum Status {
  SYNC = '已同步',
  UNDO = '未录入',
  TOBE = '待审核',
  SAVED = '已保存',
  SUBMIT = '已提交',
  NOPASS = '不通过',
  CHECKED = '通过'
}